requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Mail==0.9.0
Flask-SQLAlchemy==1.0
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.6
Werkzeug==0.9.6
blinker==1.3
gnureadline==6.3.3
ipython==2.1.0
itsdangerous==0.24
numpy==1.8.1
pyaml==14.05.7
redis==2.10.1
requests==2.3.0
scikit-learn==0.15.0
scipy==0.14.0
wsgiref==0.1.2
